Fix implicit conversions
(int)/(int) = int, which drops the fractional part. Assigning the result
to a float therefore results in loss of information.

Instead, cast the result of the first function call to a float, so that
we get a (float)/(int) operation, whose result type is float.

Coverity CID#18540

+++ eid-viewer-gui/src/main/java/be/fedict/eidviewer/gui/printing/IDPrintout.java
@@ -146,9 +146,9 @@ public class IDPrintout implements Printable,ImageObserver
 
         // calculate some sizes that need to take into account the scaling of the graphics, to avoid dragging
         // those non-intuitive "/2" further along in the code.
-        float headerHeight=coatOfArms.getHeight(this)/2;
-        float coatOfArmsWidth=coatOfArms.getWidth(this)/2;
-        float photoWidth=photo.getWidth(this)/2;
+        float headerHeight=(float)(coatOfArms.getHeight(this))/2;
+        float coatOfArmsWidth=(float)(coatOfArms.getWidth(this))/2;
+        float photoWidth=(float)(photo.getWidth(this))/2;
         float headerSpaceBetweenImages = imageableWidth - (coatOfArmsWidth + photoWidth + (SPACE_BETWEEN_ITEMS * 2));
        
         logger.finest(new ToStringBuilder(this,
